Finish implementing an underline header component. Create instructions section component for Getting Started and use it as a second use-case for underline header. Refactored sub-heading class name, which affected files outside the Pitch and Getting Started scope.
