+++ src/components/app/AppNavbar/AppNavbar.jsx
@@ -60,7 +60,7 @@ export default function AppNavbar() {
         <div className={'navbar-toolbar-main'}>
           <div className={'navbar-toolbar-content container-xxl'}>
             <div className={'navbar-left-container'}>
-              <Link className={'navbar-left-section'} href="/">
+              <Link className={'navbar-left-section'} href="/public">
                 <img src="/img/logo_opensac_black_transparent_2.png" alt="Open Sacramento logo" height="60"
                      width="200"/>
               </Link>
@@ -69,7 +69,7 @@ export default function AppNavbar() {
             <div className={'navbar-middle-container'}>
               <ul className={'navbar-middle-section'}>
                 <li className={`navbar-link ${websiteURL === '/' ? 'navbar-link-selected one' : ''}`}>
-                  <Link href="/">Home</Link>
+                  <Link href="/public">Home</Link>
                 </li>
                 <li className={`navbar-link ${websiteURL === '/about' ? 'navbar-link-selected two' : ''}`}>
                   <Link href="/about">About</Link>
