+++ src/components/ui/SidebarSection.jsx
@@ -1,0 +1,41 @@
+import Link from "next/link";
+
+/**
+ * Get started instructions section.
+ * @returns {JSX.Element}
+ */
+export default function SidebarSection({children, sectionType}) {
+  return (
+    <section className={`section-${sectionType} section-container2`}>
+      <div className={`instructions-container`}>
+        {children}
+      </div>
+
+      <div className={`sidebar-container`}>
+        <div className={`get-started`}>
+          <img src="/img/about/rectangle.png" alt="rectangle"/>
+          <p className={`get-started-anchor`}>
+            Not a volunteer yet?
+            <br/>
+            <a href="/">Get Started</a>
+          </p>
+        </div>
+        <div className={`anchors-container`}>
+          <Link href="/code-of-conduct" className={`anchor`}>Code of Conduct</Link>
+          <a
+            className={`anchor`}
+            href="https://docs.google.com/forms/d/e/1FAIpQLSfvzzvz8lCwUD8TEDiPt7ZmN34AB0dhG61590KM_scxYoZf5w/viewform"
+            target="_blank"
+            rel="noopener noreferrer"
+            aria-label="Link to new member form.">New Member Form</a>
+          <a
+            className={`anchor`}
+            href="https://www.meetup.com/code4sac/"
+            target="_blank"
+            rel="noopener noreferrer"
+            aria-label="Link to Meetup.">Meeting Link</a>
+        </div>
+      </div>
+    </section>
+  );
+};
