+++ styles/layout/section.scss
@@ -18,13 +18,3 @@
   background-color: $neutral-900;
   color: $neutral-50;
 }
-
-// Refactor this to use general section container class without light prefix.
-.general-section-light-container {
-  width: 50%;
-  position: relative;
-
-  @media (max-width: $screen-md) {
-    width: 100%;
-  }
-}
